requirements.txt
fastapi
uvicorn[standard]
ultralytics
requests
Pillow
numpy<2.0
gcsfs
google-cloud-secret-manager
google-cloud-storage
supabase
polyline
osmnx
python-dotenv
arq[redis] 
httpx
python-multipart